requirements.txt
fastapi
langchain
langchain-mistralai
langgraph
mistralai
numpy
python-dotenv
ruff